requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
python-dotenv>=1.0.0
notion-client>=2.2.1
aiohttp>=3.8.0
numpy<2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
streamlit>=1.28.0
pyyaml>=6.0